Add support for choice of value types
I am a bit dismayed that there is so much code to write for each use
of this feature. On the other hand, it is not used all that much in
the spec. On the third hand, the code in question is amenable to code
generation, so it can theoretically be implemented as a #[derive]
macro.